refactor: split src/filters into src/{filter,exception}
