+++ src/exception/problem-details.exception.spec.ts
@@ -1,8 +1,8 @@
 import { HttpAdapterHost } from '@nestjs/core';
 import { HttpStatus } from '@nestjs/common';
-import { HttpExceptionFilter } from './http-exception.filter';
-import { IProblemDetail } from './http-exception.interface';
-import { PROBLEM_CONTENT_TYPE } from './constants';
+import { HttpExceptionFilter } from '../filter/http-exception.filter';
+import { IProblemDetail } from '../filter/interfaces';
+import { PROBLEM_CONTENT_TYPE } from '../filter/constants';
 import { ProblemDetailsException } from './problem-details.exception';
 
 const mockGetResponse = jest.fn().mockImplementation(() => ({}));
