+++ src/filter/interfaces.ts
